perf(vis): lean vis x entry — in-process run (431->131ms)
The vis x file runner was paying vis's full ~295ms launcher boot (60 commands +
plugins + config) plus a second ~117ms Node spawn. Both removed:

- bin.ts is now a thin dispatcher: universal early setup (heap skip, --no-color,
  version env, compile cache) then lazy-import either the lean x path or the full
  CLI. The 60 command imports + cerebro + plugins move to cli-main.ts, so vis x
  never loads them.
- vis x runs the file IN-PROCESS via jiti (run-file.ts) — no second Node boot,
  and jiti transpiles the full TS surface + JSX with its own transpile cache.
  Bun still spawns (`bun run`).

Measured: vis x hello.ts 431ms -> 131ms (-70%), at the ~117ms Node floor. run and
all other commands are unchanged (verified run/version/help). Node floor is the
remaining anchor (a Rust launcher would be needed to beat it — out of scope).
